+++ Sources/EmojiKit/Views/EmojiScrollGrid.swift
@@ -26,7 +26,7 @@ public struct EmojiScrollGrid<ItemView: View, SectionView: View>: View {
     ///   - categories: The categories to list, by default `.all`.
     ///   - query: The search query to apply, if any.
     ///   - selection: The current grid selection, if any.
-    ///   - frequentEmojiProvider: The ``EmojiProvider`` to use, by default a ``MostRecentEmojiProvider``.
+    ///   - frequentEmojiProvider: The ``EmojiProvider`` to use, by default a ``EmojiProviders/MostRecentProvider``.
     ///   - geometryProxy: An optional geometry proxy, required to perform arrow/move-based navigation.
     ///   - action: An action to trigger when an emoji is tapped or picked.
     ///   - section: A grid section title view builder.
@@ -37,7 +37,7 @@ public struct EmojiScrollGrid<ItemView: View, SectionView: View>: View {
         categories: [EmojiCategory] = .all,
         query: String = "",
         selection: Binding<Emoji.GridSelection> = .constant(.init()),
-        frequentEmojiProvider: (any EmojiProvider)? = MostRecentEmojiProvider(),
+        frequentEmojiProvider: (any EmojiProvider)? = EmojiProviders.MostRecentProvider(),
         geometryProxy: GeometryProxy? = nil,
         action: @escaping (Emoji) -> Void = { _ in },
         @ViewBuilder section: @escaping (Emoji.GridSectionParameters) -> SectionView,